资源服务器可以进行验证 token 了

+++ map-common/src/main/java/com/alsritter/common/api/CommonResult.java
@@ -7,14 +7,14 @@ package com.alsritter.common.api;
  * @version 1.0
  **/
 public class CommonResult<T> {
-    private long code;
+    private int code;
     private String message;
     private T data;
 
     protected CommonResult() {
     }
 
-    protected CommonResult(long code, String message, T data) {
+    protected CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
@@ -100,11 +100,11 @@ public class CommonResult<T> {
         return new CommonResult<>(ResultCode.UNAUTHORIZED.getCode(), ResultCode.UNAUTHORIZED.getMessage(), data);
     }
 
-    public long getCode() {
+    public int getCode() {
         return code;
     }
 
-    public void setCode(long code) {
+    public void setCode(int code) {
         this.code = code;
     }
